# Models/Request/UpdateDepReq.cs
namespace exam.Models.Request
{
    public class UpdateDepReq
    {
        public long DepartementId { get; set; }
        public string? NameDep { get; set; }
        public string? CodeDep { get; set; }
    }
}